Add interfaces to objects in models
