+++ packages/solidarity-client/src/models/agent.ts
@@ -1,0 +1,15 @@
+import { action } from 'easy-peasy';
+import dicioAgent from '../pages/Match/Table/dicioAgent'
+
+type agentTypes = keyof typeof dicioAgent | 'default'
+
+const data: agentTypes = 'default'
+
+const agentModel = {
+  data,
+  setAgent: action((state, payload) => ({
+    data: payload
+  }))
+};
+
+export default agentModel
