--- packages/solidarity-client/src/models/agent.tsx
@@ -1,10 +1,0 @@
-import { action } from 'easy-peasy';
-
-const agentModel = {
-  data: "",
-  setAgent: action((state, payload) => ({
-    data: payload
-  }))
-};
-
-export default agentModel
